core/local/steps/producer: Merge platforms producers
  The differences between the Linux and Windows producers were so
  limited (i.e. a delay before the first scan on Linux and not using the
  `stater` module to get the file kind on both platforms but in 2
  different situations) lead us to think that they could be merged
  together, thus avoiding code duplication and creating new unit tests
  for the Windows producer.
